Seguridad terminada en clientes

+++ routes/web.php
@@ -11,20 +11,20 @@ Auth::routes();
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
 Route::group(['prefix' => 'client'], function () {
-    Route::get('/', [ClientController::class, 'index'])->name('client.index');
+    Route::get('/', [ClientController::class, 'index'])->name('client.index')->middleware(['auth', 'admin']);
 
-    Route::get('/create', [ClientController::class, 'store'])->name('client.store');
-    Route::post('/create', [ClientController::class, 'create'])->name('client.create');
+    Route::get('/create', [ClientController::class, 'store'])->name('client.store')->middleware(['auth', 'admin']);
+    Route::post('/create', [ClientController::class, 'create'])->name('client.create')->middleware(['auth', 'admin']);
 
-    Route::get('/edit/{id}', [ClientController::class, 'edit'])->name('client.edit');
-    Route::put('/update/{id}', [ClientController::class, 'update'])->name('client.update');
+    Route::get('/edit/{id}', [ClientController::class, 'edit'])->name('client.edit')->middleware(['auth', 'admin']);
+    Route::put('/update/{id}', [ClientController::class, 'update'])->name('client.update')->middleware(['auth', 'admin']);
 
-    Route::get('/editImage/{id}', [ClientController::class, 'editImage'])->name('client.editImage');
-    Route::patch('/updateImage/{id}', [ClientController::class, 'updateImage'])->name('client.updateImage');
+    Route::get('/editImage/{id}', [ClientController::class, 'editImage'])->name('client.editImage')->middleware(['auth', 'admin']);
+    Route::patch('/updateImage/{id}', [ClientController::class, 'updateImage'])->name('client.updateImage')->middleware(['auth', 'admin']);
 
-    Route::delete('/destroy/{id}', [ClientController::class, 'destroy'])->name('client.destroy');
+    Route::delete('/destroy/{id}', [ClientController::class, 'destroy'])->name('client.destroy')->middleware(['auth', 'admin']);
 
-    Route::get('/{id}', [ClientController::class, 'show'])->where('id', '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$')->name('client.show');
+    Route::get('/{id}', [ClientController::class, 'show'])->where('id', '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$')->name('client.show')->middleware(['auth', 'admin']);
 });
